+++ packages/bluesky/lib/src/extension/at_uri_extension.dart
@@ -20,8 +20,20 @@ extension AtUriExtension on core.AtUri {
   /// Returns true if this [collection] is feed like, otherwise false.
   bool get isFeedLike => collection == ids.appBskyFeedLike;
 
+  /// Returns true if this [collection] is feed generator, otherwise false.
+  bool get isFeedGenerator => collection == ids.appBskyFeedGenerator;
+
   /// Returns true if this [collection] is graph follow, otherwise false.
   bool get isGraphFollow => collection == ids.appBskyGraphFollow;
+
+  /// Returns true if this [collection] is graph block, otherwise false.
+  bool get isGraphBlock => collection == ids.appBskyGraphBlock;
+
+  /// Returns true if this [collection] is graph list, otherwise false.
+  bool get isGraphList => collection == ids.appBskyGraphList;
+
+  /// Returns true if this [collection] is graph block, otherwise false.
+  bool get isGraphListItem => collection == ids.appBskyGraphListItem;
 
   /// Returns true if this [collection] is not actor profile, otherwise false.
   bool get isNotActorProfile => !isActorProfile;
@@ -35,6 +47,18 @@ extension AtUriExtension on core.AtUri {
   /// Returns true if this [collection] is not feed like, otherwise false.
   bool get isNotFeedLike => !isFeedLike;
 
+  /// Returns true if this [collection] is not feed generator, otherwise false.
+  bool get isNotFeedGenerator => !isFeedGenerator;
+
   /// Returns true if this [collection] is not graph follow, otherwise false.
   bool get isNotGraphFollow => !isGraphFollow;
+
+  /// Returns true if this [collection] is not graph block, otherwise false.
+  bool get isNotGraphBlock => !isGraphBlock;
+
+  /// Returns true if this [collection] is not graph list, otherwise false.
+  bool get isNotGraphList => !isGraphList;
+
+  /// Returns true if this [collection] is not graph block, otherwise false.
+  bool get isNotGraphListItem => !isGraphListItem;
 }