+++ packages/bluesky/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Release Note
 
+## v0.7.3
+
+- Added callbacks in `RepoCommitAdaptor` and properties in `AtUriExtension`. ([#586](https://github.com/myConsciousness/atproto.dart/issues/586))
+  - `AtUriExtension`
+    - `isFeedGenerator`, `isNotFeedGenerator`
+    - `isGraphBlock`, `isNotGraphBlock`
+    - `isGraphList`, `isNotGraphList`
+    - `isGraphListItem`, `isNotGraphListItem`
+  - Added callbacks in `RepoCommitAdaptor`
+    - `onCreateGenerator`
+    - `onCreateBlock`
+    - `onCreateList`
+    - `onCreateListItem`
+    - `onDeleteGenerator`
+    - `onDeleteBlock`
+    - `onDeleteList`
+    - `onDeleteListItem`
+
 ## v0.7.2
 
 - A method to retrieve response data as JSON from endpoints that perform GET communication has been added. These processes do not convert to a specific model object, thus improving performance. ([#563](https://github.com/myConsciousness/atproto.dart/issues/563))